Multiple styling additions

+++ server/routes/comments.py
@@ -12,125 +12,115 @@ class Comments(Resource):
 class CommentsByID(Resource):
     def get(self, id):
         comment = get_first(Comment, 'id', id)
-        if comment:
-            return comment.to_dict(), 200
         if not comment:
             return unfound_error('Comment')
+        
+        return comment.to_dict(), 200
 
     def patch(self, id):
         comment = get_first(Comment, 'id', id)
+        current_user = get_current_user()
         data = request.get_json()
         
         if not comment:
             return unfound_error('Comment')
         
-        current_user = get_current_user()
-        if comment:
-            if comment.user_id == session.get('user_id') or current_user.admin == '1':
-                for key in data.keys():
-                    if key not in ['id', 'user_id', 'recipe_id']:
-                        setattr(comment, key, data[key])
-                        db.session.commit() 
-                        return comment.to_dict(), 201
-
+        if comment.user_id != session.get('user_id') or current_user.admin != "1":
             return unauth_error 
+        
+        comment.comment = data['comment']
+
+        db.session.commit()
+        return comment.to_dict(), 202
+
 
     def delete(self, id):
         comment = get_first(Comment, 'id', id)
+        current_user = get_current_user()
         
         if not comment:
             return unfound_error('Comment')
         
-        current_user = get_current_user()
-        if comment:
-            if comment.user_id == session.get('user_id') or current_user.admin == '1':
-                db.session.delete(comment)
-                db.session.commit()
-                return {'message':'Comment deleted'}, 204
+        if comment.user_id != session.get('user_id') or current_user.admin != '1':
+            return unauth_error
+         
+        db.session.delete(comment)
+        db.session.commit()
+        return {'message':'Comment deleted'}, 204
             
-            return unauth_error 
 
 class UserComments(Resource):
     def get(self, id):
         user = get_first(User, 'id', id)
+        current_user = get_current_user()
+        
         if not user:
             return unfound_error('User')
         
-        current_user = get_current_user()
-        if user:
-            if user.id == session.get('user_id') or current_user.admin == '1':
-                comments = [comment.to_dict() for comment in user.comments]
-                return comments, 200
-            
+        if user.id != session.get('user_id') or current_user.admin != '1':
             return unauth_error
+                
+        comments = [comment.to_dict() for comment in user.comments]
+        return comments, 200
+            
         
     
 class UserCommentsByID(Resource):
     def get(self, id, comment_id):
         user = get_first(User, 'id', id)
-        if not user:
-            return unfound_error('User')
-        
         current_user = get_current_user()
-        if user:
-            if user.id == session.get('user_id') or current_user.admin == '1':
-                comment = get_first(Comment, 'id', comment_id)
-                if not comment:
-                    return unfound_error('Comment')
-                
-                if comment.user_id is not user.id:
-                    return {'error':'This comment was not posted by this user. '}, 409
-                
-                if comment:
-                    response = comment.to_dict(), 200
-                    return response
-            
+        comment = get_first(Comment, 'id', comment_id)
+        
+        if not user or not comment:
+            return unfound_error('Item')
+        
+        if user.id != session.get('user_id') or current_user.admin != '1':
             return unauth_error
+                
+        if comment.user_id is not user.id:
+            return {'error':'This comment was not posted by this user. '}, 409
+                
+        response = comment.to_dict(), 200
+        return response
+            
     
     def patch(self, id, comment_id):
         user = get_first(User, 'id', id)
-        if not user:
-            return unfound_error('User')
-        
+        data = request.get_json()
         current_user = get_current_user()
-        if user:
-            if user.id == session.get('user_id') or current_user.admin == '1':
-                comment = get_first(Comment, 'id', comment_id)
-                if not comment:
-                    return unfound_error('Comment')
-                
-                if comment.user_id is not user.id:
-                    return {'error':'This comment was not posted by this user. '}, 409
-                
-                if comment.user_id == user.id :
-                    data = request.get_json()
-                    for key in data.keys():
-                        if key not in ['id', 'user_id', 'recipe_id']:
-                            setattr(comment, key, data[key])
-                            db.session.commit()
-                            return comment.to_dict(), 202
-
+        comment = get_first(Comment, 'id', comment_id)
+        
+        if not user or not comment:
+            return unfound_error('Item')
+        
+        if user.id != session.get('user_id') or current_user.admin != '1':
             return unauth_error  
+                
+        if comment.user_id is not user.id:
+            return {'error':'This comment was not posted by this user. '}, 409
+                
+        comment.comment = data['comment']
+        
+        db.session.commit()
+        return comment.to_dict(), 202
+
 
     def delete(self, id, comment_id):
         user = get_first(User, 'id', id)
-        if not user:
-            return unfound_error('User')
-        
         comment = get_first(Comment, 'id', comment_id)
-        
-        if not comment:
-            return unfound_error('Comment')
-        
         current_user = get_current_user()
-        if comment:
-            if comment.user_id == session.get('user_id') or current_user.admin == '1':
-                db.session.delete(comment)
-                db.session.commit()
-
-                return {'message': 'Comment deleted.'}, 204
         
-        return unauth_error 
+        if not user or not comment:
+            return unfound_error('Item')
+        
+        if comment.user_id != session.get('user_id') or current_user.admin != '1':
+            return unauth_error 
+        
+        db.session.delete(comment)
+        db.session.commit()
+
+        return {'message': 'Comment deleted.'}, 204
+        
 
 class RecipeComments(Resource):
     def get(self, id):
@@ -146,80 +136,69 @@ class RecipeComments(Resource):
         if not recipe:
             return unfound_error('Recipe')
         
-        if recipe:
-            new_comment = Comment(
-                user_id = session.get('user_id'),
-                comment = data['comment'],
-                recipe_id = id
-            )
-            db.session.add(new_comment)
-            db.session.commit()
+        new_comment = Comment(
+            user_id = session.get('user_id'),
+            comment = data['comment'],
+            recipe_id = id
+        )
+        db.session.add(new_comment)
+        db.session.commit()
 
-            return new_comment.to_dict(), 202
+        return new_comment.to_dict(), 202
 
 class RecipeCommentsByID(Resource):
     def get(self, id, comment_id):
         recipe = get_first(Recipe, 'id', id)
         comment = get_first(Comment, 'id', comment_id)
-        if not recipe:
-            return unfound_error('Recipe')
         
-        if not comment:
-            return unfound_error('Comment')
+        if not recipe or not comment:
+            return unfound_error('Item')
         
-        if recipe:
-            if comment.recipe_id == id:
-                return comment.to_dict(), 200
-            
-            else:
+        if comment.recipe_id != id:
                 return unrelated_err('Comment', 'Recipe') 
         
-
+        return comment.to_dict(), 200
+            
+        
     def patch(self, id, comment_id):
         recipe = get_first(Recipe, 'id', id)
         comment = get_first(Comment, 'id', comment_id)
         current_user = get_current_user()
-        if not recipe:
-            return unfound_error('Recipe')
+        data = request.get_json()
         
-        if recipe:
-            if not comment:
-                return unfound_error('Comment')
+        if not recipe or not comment:
+            return unfound_error('Recipe')
             
-            if comment.recipe_id is not id:
-                return unrelated_err('Comment', 'Recipe') 
+        if comment.recipe_id is not id:
+            return unrelated_err('Comment', 'Recipe') 
             
-            if comment.user_id == session.get('user_id') or current_user.admin == '1':
-                data = request.get_json()
-                for key in data.keys():
-                    if key not in ['id', 'user_id', 'recipe_id']:
-                        setattr(comment, key, data[key])
-                        db.session.commit()
-                        return comment.to_dict(), 202
-            
+        if comment.user_id != session.get('user_id') or current_user.admin != '1':
             return unauth_error
+        
+        comment.comment = data['comment']
+                        
+        db.session.commit()
+        return comment.to_dict(), 202
             
-
+            
     def delete(self, id, comment_id):
         recipe = get_first(Recipe, 'id', id)
         comment = get_first(Comment, 'id', comment_id)
         current_user = get_current_user()
-        if not recipe:
-            return unfound_error('Recipe')
         
-        if recipe:
-            if not comment:
-                return unfound_error('Comment')
+        if not recipe or not comment:
+            return unfound_error('Item')
             
-            if comment.recipe_id is not id:
+        if comment.recipe_id is not id:
                 return unrelated_err('Comment', 'Recipe') 
             
-            if comment.user_id == session.get('user_id') or current_user.admin == '1':
-                db.session.delete(comment)
-                db.session.commit()
-                return {'message':'Comment deleted'}, 204
-            
+        if comment.user_id != session.get('user_id') or current_user.admin != '1':
             return unauth_error
+                
+        db.session.delete(comment)
+        db.session.commit()
+        return {'message':'Comment deleted'}, 204
+            
 
 api.add_resource(Comments, '/api/comments')
 api.add_resource(CommentsByID, '/api/comments/<int:id>')
